% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/size.based.indicators.r
\name{CLF.f}
\alias{CLF.f}
\title{Community length frequency}
\usage{
CLF.f(species.group, codeqc)
}
\arguments{
\item{species.group}{the species group you want to be represented in the data. The PLF is generally computed 
only for demersals. Choices are "groundfish", "demersal", "pelagic". If "code", then you need to 
provide the Quebec species numerical code (codeqc).}

\item{codeqc}{the species numerical code used in Quebec region for a species. 792 is unspeciated redfish. Look at the species 
table (codeqc) for other species}
}
\description{
The Community length frequency distribution or numbers at length size spectrum. 1cm length bins are used
     where the value represents the midpoint of the length class (3= 2.5-3.5 cm). Length is generically converted to weight with W=0.1*L^3.
}
