% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.functions.r
\name{count.distinct}
\alias{count.distinct}
\title{Count distinct values in a vector}
\usage{
count.distinct(vector.name)
}
\arguments{
\item{vector.name}{name of your vector}

\item{model.type}{the kind of model to fit ("poly", "gam", "gam.adaptive","avg")}
}
\description{
Works the same as sql function count(distinct x) when using a groupby statement
}
