% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.manipulation.r
\name{datasel.f}
\alias{datasel.f}
\title{Selects sub data based on species group or a species code for further analysis}
\usage{
datasel.f(ngsl.comm.data, species.groups, species.group, codeqc)
}
\arguments{
\item{ngsl.comm.data}{ngsl community data}

\item{species.groups}{ngsl species groups with codes (codeqc)}

\item{species.group}{the species group you want to be represented in the data. The PLF is generally computed only for demersals
If "code", then you need to provide the Quebec species numerical code.}

\item{codeqc}{the species numerical code used in Quebec region for a species. 792 is unspeciated redfish}
}
\description{
Called by community indicator functions. It is unlikely that you will want to call this on its own
}
