% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.functions.r
\name{decdeg.f}
\alias{decdeg.f}
\title{Convert survey lat and long positions to decimal degress}
\usage{
decdeg.f(degminsec)
}
\arguments{
\item{degminsec}{in one number e.g. 44 38' 52", input as 443852}
}
\description{
Decimal degrees
}
