% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.functions.r
\name{gam.trend.line.f}
\alias{gam.trend.line.f}
\title{Put a smoothed gam trends line on a two dimensional graph}
\usage{
gam.trend.line.f(x, y, ...)
}
\arguments{
\item{x}{independent variable}

\item{y}{dependent variable}

\item{...}{graphical parameters accepted by the function "lines"}
}
\description{
Put a smoothed gam trends line on a two dimensional graph
}
\keyword{function}
\keyword{gam}
\keyword{helper}
