% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.functions.r
\name{map.f}
\alias{map.f}
\title{Wrapper for the PBSmapping plotMap function}
\usage{
map.f(longs = c(-74, -50), lats = c(43, 52.25),
  land.colour = "sienna3", sea.colour = "lightblue")
}
\arguments{
\item{longs}{longitude coordinates for map corners (western hempshere is -)}

\item{lats}{latitude coordinates for map corners (southern hemisphere is -)}

\item{land.colour}{colour of the land on the map}

\item{sea.colour}{colour of the sea on the map}
}
\description{
Wrapper for the PBSmapping plotMap function
}
\keyword{function}
\keyword{helper}
