% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post.analyses.plotting.r
\name{plot.CLF}
\alias{plot.CLF}
\title{Plot the community length frequency showing the slope fit line}
\usage{
\method{plot}{CLF}(CLF.object, min.len = 0, max.len = 10^6)
}
\arguments{
\item{CLF.object}{the list resulting from running the function CLF.f on the survey dataset}

\item{min.len}{the minimum length to consider in the plot and for the linear model trend}

\item{max.len}{the maximum length to consider in the plot and for the linear model trend}
}
\description{
Plots the data for all years in a matrix of plots, one year per plot.
}
