% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post.analyses.plotting.r
\name{plot.CLFab}
\alias{plot.CLFab}
\title{Plots linear model slopes over time fitted to community length frequency distribution and runs a gam smoother through them}
\usage{
\method{plot}{CLFab}(x)
}
\arguments{
\item{x}{a CLFab class object which results from fitting linear models to length frequencies with the function CLF.lm.fit}
}
\description{
A plot of slopes over time with a gam trend run through them. The GAM is from mgcv with defaults.
}
