% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/size.based.indicators.r
\name{PSS.f}
\alias{PSS.f}
\title{Predation size spectrum}
\usage{
PSS.f(species.group, codeqc, PPWR = 10, PPWR.cv = 0.1, minPPWR = 3,
  QWb = 0.75)
}
\arguments{
\item{species.group}{the species group you want to be represented in the data. The PLF is generally computed only for demersals
If "code", then you need to provide the Quebec species numerical code.}

\item{codeqc}{the species numerical code used in Quebec region for a species. 792 is unspeciated redfish. Look at the species 
table (codeqc) for other species}

\item{PPWR}{the mean predator/prey weight ratio. If 10, then the predator on average weighs 10x more than its prey}

\item{PPWR.cv}{the coefficient of variation of the predator prey weight ratio}

\item{minPPWR}{the minimum predator/prey weight ratio allowed. e.g. if 3 then the prey is never allowed to be more than 1/3 of the 
predator weight}

\item{QWb}{the exponent of the allometric relationship describing food consumption rate as a function of a fish's weight}
}
\value{
A list is returned: element 1: the year vector; element 2: the prey length vector; element 3: the prey weight vector; 
        element 4: a matrix of the predation pressure over all prey sizes (rows) by year (columns). You will likely work with 
        element 1 and 4 for plotting or creation of predation indices on particular prey size classes. The values in the matrix are 
        not scaled but this can be done easily by dividing the matrix by the maximum value of the matrix. They can be scaled annually 
        using the apply function and dividing each value in a column (year) by the maximum in that column. The whole PSS is normalised 
        by dividing by the maximum value.
}
\description{
The predator-prey weight ratio is modelled as a log-normal distribution. The energy demand of a predator is a 
        function of the predator size. This is not actually the amount of predation that has occured but a predation risk or 
        preference, i.e. given the size distribution of predators in the system, the PSS shows where they would most like to 
        take their prey. Since the index is relative, a QWa parameter is not needed as it is just a scaling parameter.
        
        The log normal density function will produce NaNs for some size ranges and warnings will be produced. This is generally 
        nothing to worry about as NaNs are converted to 0. The warnings have not be suppressed, however.
}
\references{
Duplisea, D.E. 2005. Running the gauntlet: the predation environment of small fish in the northern Gulf of St Lawrence, Canada. ICES Journal of Marine Science 62: 412-416.
}
