% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.manipulation.r
\name{q.correct.f}
\alias{q.correct.f}
\title{Determine length based catchability by fish morphotype}
\usage{
q.correct.f(L, alpha, beta, gamma)
}
\arguments{
\item{L}{length of the individual, cm}

\item{alpha}{length at q=0.5}

\item{beta}{the slope of the logistic curve}

\item{gamma}{the asymptote of the logistic curve}

\item{morpho.group}{a morpho-group designation for each species which describes the body form and preferred habitat as a means of q correcting the abundance index}
}
\description{
A logistic curve parameterised on assessment results from nGSL selectivity curves as well as literature
}
\references{
Harley, S.J. and Myers, R.A., 2001. Hierarchical Bayesian models of length-specific catchability of research trawl surveys. Canadian Journal of Fisheries and Aquatic Sciences 58: 1569-1584.
}
