% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/size.R
\name{size}
\alias{size}
\title{Sample size, power, and duration time calculation for two-arm
clinical trials with survival endpoints.}
\usage{
size(
  np = 0,
  pc = 0,
  pcratio = 0,
  k = NA,
  n = NA,
  power = NA,
  din = 0,
  dout = 0,
  loss = 0,
  sbdv = 20,
  alpha = 0.05,
  simult = 1,
  rectime = 0,
  recratio = 0,
  recrate = 0,
  ratio = 1,
  prop = 1,
  lag = 0,
  lagdout = 1,
  distpc = 0,
  distk = 0,
  distdin = 0,
  distdout = 0,
  distloss = 0,
  diratio = 0,
  doratio = 0,
  loratio = 0,
  nmin = 0,
  nmax = 20000,
  ntol = 0.001,
  npmin = 0,
  npmax = 200,
  nptol = 0.001,
  itmax = 99,
  logr = 1,
  output = 0
)
}
\arguments{
\item{np}{number of periods. The hazard rate is constant within each period.}

\item{pc}{probability of an event in the control group. \code{pc}
may be entered in two ways. First, \code{pc} may be entered as a
vector of \code{np} event probabilities. Each entry in the vector
is a conditional probability of having an event, given having
survived the previous periods. Second, \code{pc} may contain a
vector of two values. The first value is the probability of an
event and the second value is the number of periods to reach that
probability. SIZE allows multiple vectors of \code{pc} and a sample
 size or power is calculated for each vector.}

\item{pcratio}{ratio of the hazard rates for the study endpoint
over the \code{np} periods in the control group. If \code{pc} is
specified in the second way, \code{pcratio} must be specified.}

\item{k}{hazard ratio. Values of \code{k} may be entered in two ways:
(i) \code{k} = a vector of numbers if prop=0.
(ii) \code{k} = scalar if \code{prop}=1.}

\item{n}{total sample size. If the goal of the analysis is to
calculate power, \code{n} must be specified. SIZE allows multiple
values of \code{n}. Power is calculated for each \code{n}.}

\item{power}{If the goal of the analysis is to find sample sizes,
power must be specified. SIZE allows multiple values of power.
A sample size is calculated for each.}

\item{din}{the proportion of participants in the control group who
switch to the treatment group during the course of follow-up.
The default value is 0.}

\item{dout}{the proportion of participants in the treatment group
who switch to the control group during the course of follow-up.
The default value is 0.}

\item{loss}{the proportion of participants lost to follow-up because
of censoring, loss of contact, or competing risks.}

\item{sbdv}{number of intervals to be divided in each period.
Within a given period, the probability of an event is assumed the
same across intervals of  equal length and equals
 \eqn{1 - (1 - x)^{1/sbdv}}, where \eqn{x} is the probability of
 an event in a period. If \code{lag} is not specified, then the
 default of \code{sbdv} is 20.  If \code{lag} is specified,
 the default is 4. The product of \code{lag} and
\code{sbdv} determines the number of intermediate states for the
event rate in the treatment group. For example, if \code{lag} = 5 and
\code{sbdv} = 4, then there are 5 X 4 = 20 intermediate states
between the control and treatment groups. The intermediate states
account for the lag in treatment effect.}

\item{alpha}{type I error for two-tailed alternatives. The default value is 0.05.}

\item{simult}{0 or 1. If \code{simult} = 1, then participants
enter the study simultaneously. If \code{simult} = 0, participants
enter the study according to the accrual pattern specified in
\code{rectime} and \code{recratio}.}

\item{rectime}{a vector of accrual time. If \code{simult} equals 0,
\code{rectime} must be specified.}

\item{recratio}{a vector of relative accrual rates over the
accrual times. The length of $recratio$ and the length of
\code{rectime} must be equal.}

\item{recrate}{the recruitment rate in each period. If both sample
size and power are specified and the goal is to solve for the
duration time, then \code{recrate} must be specified.}

\item{ratio}{two numbers referring to the
ratio of the sample size in the control group to that in the
treatment group. For example, if the allocation ratio is 2:3,
then \code{ratio}  = c(2, 3).}

\item{prop}{0 or 1. If \code{prop} = 1, then the hazards are
proportional and \code{k} remains the same for the duration of the study.
If \code{prop} = 0, then the hazard are nonproportional and
\code{k} must be a vector of np entries.}

\item{lag}{lag time, the number of periods to achieve the full
treatment effect. It adjusts the event rate for crossovers under
the non-proportional hazards model. If \code{lag} is not specified,
then participants return to the efficacy level comparable to the
level in the opposite arm immediately after crossover.}

\item{lagdout}{lagged effect of dropout. It equals 1 if crossovers
from the treatment group return to the control group level in the
same fashion as they reach the treatment effect level and equals
0 if the treatment effect vanishes immediately after crossover.
The default value is 1.}

\item{distpc}{a discrete prior probability distribution for $pc$.
Each vector of \code{pc} is considered a mass point and is assigned a
probability mass. The number of masses in \code{distpc} must be equal
to the number of vectors in \code{pc}. The sum of values of \code{distpc} is 1.}

\item{distk}{a discrete prior probability distribution for the
hazard ratio, \code{k}. Each \code{k} value is considered a mass point and
is assigned a probability mass from \code{distk}. The number of masses
in \code{distk} must be equal to the number of entries in \code{k}.
The sum of values of \code{distk} is 1.}

\item{distdin}{a discrete prior probability distribution for \code{din}.
The sum of values of \code{distdin} is 1.}

\item{distdout}{a discrete prior probability distribution for \code{dout}.
The sum of values of \code{distdout} is 1.}

\item{distloss}{a discrete prior distribution for \code{loss}.
The sum of values of \code{distloss} is 1.}

\item{diratio}{ratio of the hazard rates for crossovers from the
control to the treatment group.}

\item{doratio}{ratio of the hazard rates for crossovers from the
treatment to the control group.}

\item{loratio}{ratio of the hazard rates for loss.}

\item{nmin, nmax, ntol}{minimum and maximum of the sample size to be used
in the iterative bisection method to search for the sample size
such that the calculated power is equal to or grater than the
specified power within the relative tolerance, \code{ntol}.
Default  = 0, 20000, 0.001.}

\item{npmin, npmax, nptol}{minimum and maximum of the duration time
to be used in the iterative bisection method to search for the
duration time such that the calculated power is equal to or
greater than the specified power within the relative tolerance,
\code{nptol}. Default  = 0, 200, 0.001.}

\item{itmax}{The maximum number of iterations.}

\item{logr}{indicator of the test statistic. If \code{logr} = 0,
then the binomial formula is used. If \code{logr} = 1, then the log-rank
test is used. If \code{logr} = a vector of \code{np} weights, then the
weighted log-rank test is used. The default value is 1.}

\item{output}{0 or 1. If \code{output} = 1, then all the parameter values
are printed. The default value is 0.}
}
\value{
The output list includes the following elements:
\itemize{
  \item \code{alpha} - type I error for two-tailed alternatives. The default value is 0.05.
  \item \code{power} - the probability of making a correct decision (to reject the null hypothesis) when the null hypothesis is false.
  \item \code{k} - hazard ratio.
  \item \code{din} - the proportion of participants in the control group who switch to the treatment group during the course of follow-up.
  \item \code{dout} - the proportion of participants in the treatment group who switch to the control group during the course of follow-up.
  \item \code{loss} - the proportion of participants lost to follow-up because of censoring, loss of contact, or competing risks.
  \item \code{pc} - probability of an event in the control group.
  \item \code{pe} - probability of an event in the experimental group.
  \item \code{n} - total sample size.
  \item \code{event} - number of event.
  \item \code{duration} - duration time.
  }
}
\description{
Sample size, power or duration of the study are
calculated for survival endpoints with time-dependent rates of
event, crossover, and loss to follow-up. The calculation allows
for the specification of a wide range of complexities commonly
occurring in clinical trials including nonproportional hazards,
lag in treatment effect, and uncertainties in treatment benefit.
It was based on the method of Lakatos (1986, 1988) and previously
implemented in SAS.
}
\examples{
## Examples with parameter entered with two values

## Sample size
size(
np=8,
pc=c(0.30, 8),
pcratio=rep(1,8),
k=c(0.65,0.35),
distk=c(0.6,0.4),
power=0.8)

## Power
size(
np=8,
pc=c(0.30, 8),
pcratio=rep(1,8),
k=c(0.65,0.35),
distk=c(0.6,0.4),
n=800)

## Duration time
size(
np=4,
pc=c(0.30,4),
pcratio=c(1,2,2,2),
k=0.65,
power=0.8,
n=800,
din=c(0.10,4),
diratio=c(1,1,1,1),
dout=c(0.10,4),
doratio=c(1,1,2,2),
loss=c(0.10,4),
loratio=c(1,1,1,1),
npmin=1,
npmax=20,
recrate=200)

## Examples with parameter entered with vector

## Sample size
size(np=4,
pc=c(c(0.35,0.3,0.25,0.2),c(0.3,0.25,0.2,0.15)),
distpc=c(0.6, 0.4),
k=c(1,0.82,0.65,0.65,1,0.72,0.75,0.45),
prop=0,
distk=c(0.6, 0.4),
power=0.8,
din=c(rep(0.005, 4), rep(0.002, 4)),
distdin=c(0.6, 0.4),
dout=c(rep(0.005, 4), rep(0.002, 4)),
distdout=c(0.6, 0.4),
loss=c(rep(0.15, 4), rep(0.05, 4)),
distloss=c(0.6, 0.4),
logr=1)


## Power
size(
np=4,
pc=c(c(0.35,0.3,0.25,0.2),c(0.3,0.25,0.2,0.15)),
distpc=c(0.6, 0.4),
k=c(1,0.82,0.65,0.65,1,0.72,0.75,0.45),
prop=0,
distk=c(0.6, 0.4),
n=800,
din=c(rep(0.005, 4), rep(0.002, 4)),
distdin=c(0.6, 0.4),
dout=c(rep(0.005, 4), rep(0.002, 4)),
distdout=c(0.6, 0.4),
loss=c(rep(0.15, 4), rep(0.05, 4)),
distloss=c(0.6, 0.4),
logr=1)


## Duration time
size(
np=4,
pc=c(0.35,0.3,0.25,0.2),
k=c(1,0.82,0.65,0.65),
prop=0,
power=0.8,
n=800,
din=rep(0.005, 4),
dout=rep(0.005, 4),
loss=rep(0.15, 4),
npmin=1,
npmax=20,
recrate=200)

}
\references{
Lakatos, Edward. "Sample size determination in clinical trials with
time-dependent rates of losses an noncomplinance."
Controlled Clinical Trials 7:189-199,1986.

Lakatos, Edward. "Sample sizes based on the log-rank statistic in
complex clinical trials." Biometrics 44:229-241, 1988.

Shih, Joanna H."Sample Size Calculation for Complex Clinical Trials
with Survival Endpoints", Controlled Clinical Trials, 16:395-407,
1995.
}
\author{
Qihu Zhang & Joanna H. Shih
}
