% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post.analyses.plotting.r
\name{slope.slider.f}
\alias{slope.slider.f}
\title{A wrapper to use sliders to examine impact of length cutoffs on the trend in linear slopes over time}
\usage{
slope.slider.f(CLF.object, min.len = 0, max.len = 10^6)
}
\arguments{
\item{min.len}{the minimum length to consider in the plot and for the linear model trend}

\item{max.len}{the maximum length to consider in the plot and for the linear model trend}

\item{x}{a CLFab class object which results from fitting linear models to length frequencies with the function CLF.lm.fit}
}
\description{
A plot of slopes over time with a gam trend run through them. The GAM is from mgcv with defaults.
}
\examples{
library(manipulate)
        tmp= CLF.f("all")
        manipulate(slope.slider.f(tmp, min.len=x.min, max.len=x.max), x.min= slider(1,35),x.max= slider(70,150))
}
