% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/size.based.indicators.r
\name{SS.f}
\alias{SS.f}
\title{Numbers at log2 weight size spectrum}
\usage{
SS.f(species.group, codeqc)
}
\arguments{
\item{species.group}{the species group you want to be represented in the data. The PLF is generally computed 
only for demersals. Choices are "groundfish", "demersal", "pelagic". If "code", then you need to 
provide the Quebec species numerical code (codeqc).}

\item{codeqc}{the species numerical code used in Quebec region for a species. 792 is unspeciated redfish. Look at the species 
table (codeqc) for other species}
}
\description{
The numbers an biomass at log2 weight class size spectrum for the community components you select. 
        This is the classic size spectrum of Sheldon & Dickie
}
\references{
Duplisea, D.E. and Castonguay, M. 2006. Comparison and utility of different size-based metrics of fish communities for detecting fishery impacts. Canadian Journal of Fisheries and Aquatic Sciences 63: 810-820.
}
