% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/size.based.indicators.r
\name{PLF.f}
\alias{PLF.f}
\title{Compute the proportion of large fish from the survey data}
\usage{
PLF.f(cutoff = 30, species.group = "demersal", codeqc = 792)
}
\arguments{
\item{cutoff}{the size (cm) marking the ditinction between large and small fish. Large fish are >= to this size.}

\item{species.group}{the species group you want to be represented in the data. The PLF is generally computed only for demersals
If "code", then you need to provide the Quebec species numerical code.}

\item{codeqc}{the species numerical code used in Quebec region for a species. 792 is unspeciated redfish. Look at the species 
table (codeqc) for other species}
}
\description{
The proporption of large fish indicator by abundance.
}
